Stop passing --check-certificate=false
The original idea was: If portage is verifying checksums anyway,
we can tolerate bad certificates in favor of more user convenience.

However, we may just be fetching the file for the very first time
(in case of an Ebuild developer) and would not notice a MITM attack.
So in that case, it makes a dangerous default.